! Pay with promotion and sale

+++ BackEnd/FoodOrder.Core/Models/Order.cs
@@ -14,7 +14,7 @@ namespace FoodOrder.Core.Models
         public DateTime? DatePaid { get; set; }
         public int OrderStatusID { get; set; }
         public int? PromotionID { get; set; }
-        public decimal? PromotionAmount { get; set; }
+        public double? PromotionAmount { get; set; }
 
         public string AddressString { get; set; }
         public string AddressName { get; set; }
