+++ BackEnd/FoodOrder.Core/Models/Promotion.cs
@@ -16,8 +16,8 @@ namespace FoodOrder.Core.Models
         public int UseTimes { get; set; } // So lan co the su dung
         public float Percent { get; set; }
         public bool Enabled { get; set; }
-        public int Max { get; set; }
-        public int MinPrice { get; set; }
+        public double Max { get; set; }
+        public double MinPrice { get; set; }
         public bool IsGlobal { get; set; }
         public int Priority { get; set; }
